Modify the way to find WIF for ScriptPubKey to use PubKeyHash for tx generate
The current logic assumes the WIF and the ScriptPubKey are tightly bound
in the input format. However, the ScriptPubKey includes PubKeyHash, and
it's possible to find WIF that has matching PubKey.

This commit will change the logic for `tx generate` subcommand to use
PubKeyHash, and also modify the input format to follow the change.